refactor: [#248] Move Trystero handlers to their own file
